Update README.md with some research notes
The example code was surprisingly easy to build and run, getting
instantaneous results. This prompted me to take a step back and
do some research to see what prior art there is and what other options
there are in Win32. This lead to some interesting results, noted down in
the README.